feat: Deprecate codegen feat: Remove codegen
