--- codegen/.snapshots/TestGenerateTemplate-should_add_comma_between_relations
@@ -1,13 +1,0 @@
-{
-		Name:         "with relations",
-		Columns: []schema.Column{
-
-		},
-
-		Relations: []*schema.Table{
-relation1,
-relation2,
-
-		},
-
-}
