--- codegen/.snapshots/TestGenerateTemplate-should_add_ignore_in_tests_to_columns
@@ -1,12 +1,0 @@
-{
-		Name:         "with relations",
-		Columns: []schema.Column{
-{
-  Name:        "ignore_in_tests",
-  Type:        schema.TypeString,
-  IgnoreInTests: true,
-},
-
-		},
-
-}
